Refactor ESLint configuration to modularize rules
- Moved ESLint rules into separate files for better organization and maintainability
- Updated main ESLint source file to import rules from new modules
- Improved clarity and structure of the ESLint configuration
