+++ packages/eslint-config/rules/ignored-folders.ts
@@ -1,0 +1,14 @@
+import { type Linter } from "eslint";
+
+export const ignoredFolders = {
+  ignores: [
+    "**/.react-router/**",
+    "**/node_modules/**",
+    "pnpm-lock.yaml",
+    ".vscode/*",
+    "**/.vercel/*",
+    "**/.astro/*",
+    "**/.turbo/*",
+    "**/build/*",
+  ],
+} satisfies Linter.Config;
