+++ packages/eslint-config/rules/package-json-rules.ts
@@ -1,0 +1,8 @@
+import { type Linter } from "eslint";
+
+export const packageJsonRules = {
+  files: ["**/package.json"],
+  rules: {
+    "jsonc/sort-keys": "off",
+  },
+} satisfies Linter.Config;
