+++ packages/eslint-config/rules/storybook-rules.ts
@@ -1,0 +1,9 @@
+import { type Linter } from "eslint";
+
+export const storybookRules = {
+  files: ["**/*.stories.ts", "**/*.stories.tsx"],
+  rules: {
+    "canonical/filename-match-exported": "off",
+  },
+  ignores: ["**/storybook-static/**/*"],
+} satisfies Linter.Config;
